Verify more parts of the build in Travis

+++ scripts/build_pkg.sh
@@ -3,6 +3,9 @@
 #
 # Build and package fwup
 #
+# Inputs:
+#     SKIP_PACKAGE - set to "true" to skip the package making step
+#
 # This script creates a static build of fwup to avoid dependency issues
 # with libconfuse and libsodium on various Linux distributions. The result
 # is a self-contained .deb and .rpm that should work on any Linux (assuming
@@ -18,6 +21,10 @@ DEPS_INSTALL_DIR=$DEPS_DIR/usr
 FWUP_INSTALL_DIR=$BUILD_DIR/fwup-installed/usr
 
 MAKE_FLAGS=-j8
+
+if [[ $(uname -s) = "Darwin" ]]; then
+    EXTRA_LDFLAGS="-L/usr/local/lib -lintl"
+fi
 
 # Initial sanity checks
 if [[ ! -e $BASE_DIR/configure ]]; then
@@ -42,7 +49,8 @@ pushd $BUILD_DIR
 # Build fwup (symlink now, since out-of-tree fwup build is broke)
 ln -sf $BASE_DIR $BUILD_DIR/fwup
 pushd fwup
-LDFLAGS=-L$DEPS_INSTALL_DIR/lib CPPFLAGS=-I$DEPS_INSTALL_DIR/include ./configure --prefix=$FWUP_INSTALL_DIR --enable-shared=no
+LDFLAGS="-L$DEPS_INSTALL_DIR/lib $EXTRA_LDFLAGS" CPPFLAGS=-I$DEPS_INSTALL_DIR/include ./configure --prefix=$FWUP_INSTALL_DIR --enable-shared=no || cat config.log
+cat config.log
 make clean
 make $MAKE_FLAGS
 
@@ -71,8 +79,13 @@ popd
 popd
 
 # Package fwup
-FWUP_VERSION=$($FWUP_INSTALL_DIR/bin/fwup --version)
-rm -f fwup_*.deb fwup-*.rpm
-fpm -s dir -t deb -v $FWUP_VERSION -n fwup -C $FWUP_INSTALL_DIR/..
-fpm -s dir -t rpm -v $FWUP_VERSION -n fwup -C $FWUP_INSTALL_DIR/..
+if [[ "$SKIP_PACKAGE" != "true" ]]; then
+    FWUP_VERSION=$($FWUP_INSTALL_DIR/bin/fwup --version)
+    rm -f fwup_*.deb fwup-*.rpm
+    fpm -s dir -t deb -v $FWUP_VERSION -n fwup -C $FWUP_INSTALL_DIR/..
+    fpm -s dir -t rpm -v $FWUP_VERSION -n fwup -C $FWUP_INSTALL_DIR/..
+else
+    echo "Static build was successful, but skipped creating packages."
+    echo "The fwup installation is in $FWUP_INSTALL_DIR."
+fi
